feat(rtl): add right-to-left (RTL) support to CSS
This commit introduces RTL (Right-to-Left) support to the website's CSS. It includes adjustments for text alignment, navigation, dropdown menus, flex layouts, and other elements to ensure proper rendering in RTL languages. This enhancement improves the user experience for users who read and interact with content in RTL languages such as Persian and Arabic.