Set up and Export the Persian Expansion Pack PDF
I set up XeLatex for Persian. In the Simple PDF Generation, it is mentioned we had to use the BNazanin font and put the font name within " ". But I have tried to do that multiple times but there was no B Nazanin for XeLatex after some investigations. Then I explored and found some fonts which existed in MiKTeX, and chose one that was close to the trend in Iran. If anyone wants to change it, they must read the documentation of XeLatex for Persians in MiKTex, then copy and paste the font name without " ". They will see it works.